1. fix a complie bug 2. remove a bug which when stop in windows runing.
Former-commit-id: 5d4f144134a310dde603c45af4b6da90931dc4a4 [formerly e478d78427022f0342ede27e2c40acd48f77907f]
Former-commit-id: de7c9682fb4deddcf9fb23ea6a2fdb121e83636a

+++ src/rpc/rpctx.cpp
@@ -131,7 +131,7 @@ Object GetTxDetailJSON(const uint256& txhash) {
 					obj = TxToJSON(pBaseTx.get());
 					obj.push_back(Pair("blockhash", header.GetHash().GetHex()));
 					obj.push_back(Pair("confirmHeight", (int) header.nHeight));
-					obj.push_back(Pair("confirmedtime",header.nTime));
+					obj.push_back(Pair("confirmedtime",(int)header.nTime));
 				} catch (std::exception &e) {
 					throw runtime_error(tfm::format("%s : Deserialize or I/O error - %s", __func__, e.what()).c_str());
 				}
